✔️ Game States / Entire Menu (no values and reset)

+++ Misty/src/entities/Entity.java
@@ -22,9 +22,10 @@ public abstract class Entity {
 		// For debugging the hitbox
 		g.setColor(Color.PINK);
 		g.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
+
 	}
 
-	protected void initHitbox(float x, float y, float width, float height) {
+	protected void initHitbox(float x, float y, int width, int height) {
 		hitbox = new Rectangle2D.Float(x, y, width, height);
 	}
 
@@ -36,4 +37,5 @@ public abstract class Entity {
 	public Rectangle2D.Float getHitbox() {
 		return hitbox;
 	}
+
 }
